Regen reference docs to ci
