+++ CONTRIBUTING.md
@@ -136,6 +136,14 @@ reviewed, checked-in versions. If your PR would change the generated output,
 run this command to regenerate any samples and check those files in with
 your PR. Carefully review whether the changes are intentional.
 
+## Regenerate Reference Docs
+
+```bash
+rush regen-docs
+```
+
+PR validation will ensure that reference docs are up to date.
+
 # Using VS Code
 
 ## Recommended extensions
