+++ packages/ref-doc/src/cli.ts
@@ -1,8 +1,6 @@
-import { compile, joinPaths, NodeHost, normalizePath } from "@cadl-lang/compiler";
-import { mkdir, writeFile } from "fs/promises";
+import { normalizePath } from "@cadl-lang/compiler";
 import { resolve } from "path";
-import { renderToDocusaurusMarkdown } from "./emitters/docusaurus.js";
-import { extractRefDocs } from "./extractor.js";
+import { generateDocs } from "./index.js";
 
 main().catch((e) => {
   console.error(e);
@@ -29,13 +27,5 @@ async function main() {
     resolvedOutputDir,
   });
 
-  const program = await compile(NodeHost, main, {
-    parseOptions: { comments: true, docs: true },
-  });
-  const refDoc = extractRefDocs(program, namespaces);
-  const files = renderToDocusaurusMarkdown(refDoc);
-  await mkdir(resolvedOutputDir, { recursive: true });
-  for (const [name, content] of Object.entries(files)) {
-    writeFile(joinPaths(resolvedOutputDir, name), content);
-  }
+  generateDocs(main, namespaces, resolvedOutputDir);
 }
