+++ packages/ref-doc/src/index.ts
@@ -1,1 +1,2 @@
+export * from "./experimental.js";
 export * from "./extractor.js";
